feat: extracted chain graph as a separate component
