+++ src/pages/ChainGraph.tsx
@@ -1,20 +1,8 @@
-import {
-  Background,
-  BackgroundVariant,
-  Edge,
-  type KeyCode,
-  MiniMap,
-  Node,
-  OnSelectionChangeFunc,
-  ReactFlow,
-  ReactFlowProvider,
-  useOnSelectionChange,
-} from "@xyflow/react";
+import { Node, ReactFlowProvider } from "@xyflow/react";
 
 import "@xyflow/react/dist/style.css";
 import "../styles/reactflow-theme.css";
 import React, {
-  MouseEvent,
   useCallback,
   useContext,
   useEffect,
@@ -24,27 +12,17 @@ import React, {
 } from "react";
 import { ElementsLibrarySidebar } from "../components/elements_library/ElementsLibrarySidebar.tsx";
 import { DnDProvider } from "../components/DndContext.tsx";
-import { Button, Flex, Modal } from "antd";
-import { useNavigate } from "react-router";
+import { Button, Flex } from "antd";
 import { useParams } from "react-router-dom";
 import { useRegisterChainHeaderActions } from "./ChainHeaderActionsContext.tsx";
-import { ChainGraphViewControls } from "../components/graph/ChainGraphViewControls.tsx";
 import {
-  ElementFocus,
   ElementFocusContext,
   type FitViewToElementIdFn,
 } from "../components/graph/ElementFocus.tsx";
 import { useModalsContext } from "../Modals.tsx";
-import { ChainElementModification } from "../components/modal/chain_element/ChainElementModification.tsx";
 import styles from "./ChainGraph.module.css";
 import controlStyles from "../components/graph/ChainGraphViewControls.module.css";
-import {
-  LibraryProvider,
-  useLibraryContext,
-} from "../components/LibraryContext.tsx";
-
-import { useChainGraph } from "../hooks/graph/useChainGraph.tsx";
-import { ElkDirectionContextProvider } from "./ElkDirectionContext.tsx";
+import { LibraryProvider } from "../components/LibraryContext.tsx";
 import { PageWithRightPanel } from "./PageWithRightPanel.tsx";
 import { SaveAndDeploy } from "../components/modal/SaveAndDeploy.tsx";
 import {
@@ -57,21 +35,8 @@ import { api } from "../api/api.ts";
 import { useNotificationService } from "../hooks/useNotificationService.tsx";
 import { SequenceDiagram } from "../components/modal/SequenceDiagram.tsx";
 import { ChainContext } from "./ChainPage.tsx";
-import {
-  ChainGraphNode,
-  ChainGraphNodeData,
-  nodeTypes,
-  OnDeleteEvent,
-} from "../components/graph/nodes/ChainGraphNodeTypes.ts";
 import { isVsCode } from "../api/rest/vscodeExtensionApi.ts";
-import ContextMenu from "../components/graph/ContextMenu.tsx";
 import { PanelResizeHandle } from "../components/PanelResizeHandle.tsx";
-import {
-  getElementColor,
-  isSwimlanesOnly,
-  nonEmptyContainerExists,
-  sanitizeEdge,
-} from "../misc/chain-graph-utils.ts";
 import {
   ExportChainOptions,
   ExportChains,
@@ -84,35 +49,15 @@ import { Require } from "../permissions/Require.tsx";
 import { ProtectedButton } from "../permissions/ProtectedButton.tsx";
 import { usePermissions } from "../permissions/usePermissions.tsx";
 import { hasPermissions } from "../permissions/funcs.ts";
-import { getSwimlaneBorderColor } from "../components/graph/nodes/SwimlaneNode.tsx";
-import { useContextMenu } from "../hooks/graph/useContextMenu.tsx";
 import { useGenerateDds } from "../hooks/useGenerateDds.tsx";
 import { OverridableIcon } from "../icons/IconProvider.tsx";
-
-const readTheme = () => {
-  if (typeof document === "undefined") return "light";
-  const candidates: (globalThis.Element | null)[] = [
-    document.documentElement,
-    document.body,
-    document.querySelector(".vscode-webview"),
-  ];
-
-  for (const target of candidates) {
-    if (!target) continue;
-    const value = target.getAttribute("data-theme");
-    if (value !== null && typeof value === "string") return value;
-  }
-
-  if (document.body?.classList.contains("vscode-dark")) {
-    return "dark";
-  }
-
-  if (document.body?.classList.contains("vscode-high-contrast")) {
-    return "high-contrast";
-  }
-
-  return "light";
-};
+import { ChainGraphView } from "./ChainGraphView.tsx";
+import {
+  ChainGraphNode,
+  ChainGraphNodeData,
+} from "../components/graph/nodes/ChainGraphNodeTypes.ts";
+import { ChainElementModification } from "../components/modal/chain_element/ChainElementModification.tsx";
+import { useNavigate } from "react-router";
 
 const MIN_PANEL_WIDTH = 230;
 const MAX_PANEL_WIDTH = 500;
@@ -126,60 +71,20 @@ const ChainGraphInner: React.FC = () => {
   const { chainId, elementId } = useParams<string>();
   const chainContext = useContext(ChainContext);
   const { showModal } = useModalsContext();
-  const reactFlowWrapper = useRef(null);
-  const [isPageLoaded, setIsPageLoaded] = useState(false);
-  const [currentTheme, setCurrentTheme] = useState<string>(() => readTheme());
   const [leftPanelWidth, setLeftPanelWidth] = useState(
     DEFAULT_LEFT_PANEL_WIDTH,
   );
   const [rightPanelWidth, setRightPanelWidth] = useState(
     DEFAULT_RIGHT_PANEL_WIDTH,
   );
-  const currentThemeRef = useRef(currentTheme);
-  const navigate = useNavigate();
   const notificationService = useNotificationService();
-  const { isLibraryLoading, libraryElements } = useLibraryContext();
   const permissions = usePermissions();
   const [readOnly, setReadOnly] = useState<boolean>(false);
-  const deleteKeyCode = useMemo<KeyCode | null>(
-    () => (readOnly || elementId ? null : ["Backspace", "Delete"]),
-    [elementId, readOnly],
-  );
-  const [selectedByRightClick, setSelectedByRightClick] =
-    useState<boolean>(false);
+  const navigate = useNavigate();
 
   useEffect(() => {
     setReadOnly(!hasPermissions(permissions, { chain: ["update"] }));
   }, [permissions]);
-
-  const {
-    nodes,
-    setNodes,
-    edges,
-    setEdges,
-    decorativeEdges,
-    onConnect,
-    onDragOver,
-    onDrop,
-    onDelete,
-    onEdgesChange,
-    onNodesChange,
-    onNodeDragStart,
-    onNodeDrag,
-    onNodeDragStop,
-    direction,
-    toggleDirection,
-    updateNodeData,
-    isLoading,
-    expandAllContainers,
-    collapseAllContainers,
-    structureChanged,
-  } = useChainGraph();
-
-  const renderEdges = useMemo(
-    () => [...edges, ...decorativeEdges].map(sanitizeEdge),
-    [edges, decorativeEdges],
-  );
 
   const [leftPanelVisible, setLeftPanelVisible] = useState<boolean>(true);
   const [rightPanelVisible, setRightPanelVisible] = useState<boolean>(false);
@@ -187,32 +92,27 @@ const ChainGraphInner: React.FC = () => {
   const fitViewToElementIdRef = useRef<FitViewToElementIdFn | null>(null);
   const { showGenerateDdsModal } = useGenerateDds();
 
-  const handleElementUpdated = useCallback(
-    (element: Element, node: ChainGraphNode) => {
-      updateNodeData(element, node);
-      void chainContext?.refresh();
-    },
-    [updateNodeData, chainContext],
-  );
-
   const clearElementPath = useCallback(() => {
-    void navigate(`/chains/${chainId}/graph`);
-  }, [chainId, navigate]);
+    void navigate(`/chains/${chainContext?.chain?.id}/graph`);
+  }, [chainContext, navigate]);
 
   const setElementPath = useCallback(
     (newElementId: string) => {
-      void navigate(`/chains/${chainId}/graph/${newElementId}`);
+      void navigate(`/chains/${chainContext?.chain?.id}/graph/${newElementId}`);
     },
-    [chainId, navigate],
+    [chainContext, navigate],
   );
 
   const openElementModal = useCallback(
-    (node?: Node<ChainGraphNodeData>) => {
-      if (!node?.data.elementType) return;
-      const modalId = `chain-element-${node.id}`;
+    (
+      node: Node<ChainGraphNodeData> | undefined,
+      updateCallback: (element: Element, node: ChainGraphNode) => void,
+    ) => {
+      if (!node || !node?.data.elementType) return;
       if (elementId !== node.id) {
         setElementPath(node.id);
       }
+      const modalId = `chain-element-${node.id}`;
       showModal({
         id: modalId,
         component: (
@@ -221,128 +121,18 @@ const ChainGraphInner: React.FC = () => {
               node={node}
               chainId={chainId!}
               elementId={node.id}
-              onSubmit={handleElementUpdated}
+              onSubmit={(e, n) => {
+                updateCallback(e, n);
+                clearElementPath();
+              }}
               onClose={clearElementPath}
             />
           </ChainContext.Provider>
         ),
       });
     },
-    [
-      chainContext,
-      chainId,
-      elementId,
-      clearElementPath,
-      setElementPath,
-      showModal,
-      handleElementUpdated,
-    ],
+    [chainContext, chainId, showModal],
   );
-
-  const handleDelete = useCallback(
-    async (changes: OnDeleteEvent) => {
-      if (
-        changes.nodes.length > 0 &&
-        (await nonEmptyContainerExists(changes.nodes as ChainGraphNode[])) &&
-        !isSwimlanesOnly(changes.nodes as ChainGraphNode[])
-      ) {
-        Modal.confirm({
-          title: "Delete Container",
-          content:
-            "This container element is not empty. Are you sure you want to delete it? All its content will be also deleted.",
-          onOk: () => void onDelete(changes),
-        });
-      } else {
-        void onDelete(changes);
-      }
-    },
-    [onDelete],
-  );
-
-  const { menu, closeMenu, onContextMenuCall } = useContextMenu(
-    handleDelete,
-    openElementModal,
-    nodes,
-    setNodes,
-    edges,
-    setEdges,
-    structureChanged,
-    chainId,
-    chainContext?.refresh,
-  );
-
-  const onNodeDoubleClick = (
-    _event: MouseEvent,
-    node: Node<ChainGraphNodeData>,
-  ) => {
-    openElementModal(node);
-  };
-
-  const handleSelectionChange = useCallback<
-    OnSelectionChangeFunc<Node<ChainGraphNodeData>, Edge>
-  >(() => {
-    if (selectedByRightClick) {
-      setSelectedByRightClick(false);
-    } else {
-      closeMenu();
-    }
-  }, [closeMenu, selectedByRightClick]);
-
-  useOnSelectionChange<Node<ChainGraphNodeData>, Edge>({
-    onChange: handleSelectionChange,
-  });
-
-  const onContextMenu = (event: MouseEvent) => {
-    const elements: Node<ChainGraphNodeData>[] = [];
-
-    setSelectedByRightClick(true);
-    setNodes((prev) =>
-      prev.map((prevNode) => ({
-        ...prevNode,
-        selected: false,
-      })),
-    );
-
-    (
-      onContextMenuCall as (
-        event: MouseEvent,
-        elements: Node<ChainGraphNodeData>[],
-      ) => void
-    )(event, elements);
-  };
-
-  const onNodeContextMenu = (
-    event: MouseEvent,
-    node: Node<ChainGraphNodeData>,
-  ) => {
-    event.stopPropagation();
-    const elements: Node<ChainGraphNodeData>[] = [];
-
-    const selectedBeforeRightClick = nodes.filter((n) => n.selected);
-
-    if (
-      selectedBeforeRightClick.some(
-        (selectedNode) => selectedNode.id === node.id,
-      )
-    ) {
-      elements.push(...selectedBeforeRightClick);
-    } else {
-      setSelectedByRightClick(true);
-      setNodes((prev) =>
-        prev.map((prevNode) => ({
-          ...prevNode,
-          selected: prevNode.id === node.id,
-        })),
-      );
-      elements.push(node);
-    }
-    (
-      onContextMenuCall as (
-        event: MouseEvent,
-        elements: Node<ChainGraphNodeData>[],
-      ) => void
-    )(event, elements);
-  };
 
   const saveAndDeploy = useCallback(
     async (domains: Domain[]) => {
@@ -506,129 +296,6 @@ const ChainGraphInner: React.FC = () => {
   );
   useRegisterChainHeaderActions(headerActions, [chainId]);
 
-  useEffect(() => {
-    if (!isPageLoaded && !isLoading && !isLibraryLoading && nodes?.length) {
-      setIsPageLoaded(true);
-      if (elementId) {
-        const targetNode = nodes.find((node) => node.id === elementId);
-        if (targetNode) openElementModal(targetNode);
-      }
-    }
-  }, [
-    elementId,
-    isLoading,
-    isLibraryLoading,
-    isPageLoaded,
-    nodes,
-    openElementModal,
-  ]);
-
-  useEffect(() => {
-    currentThemeRef.current = currentTheme;
-  }, [currentTheme]);
-
-  useEffect(() => {
-    if (typeof document === "undefined") {
-      return;
-    }
-
-    const updateTheme = () => {
-      const theme = readTheme();
-      if (!theme) return;
-      if (currentThemeRef.current === theme) {
-        if (document.body.dataset.theme !== theme) {
-          document.body.dataset.theme = theme;
-        }
-        return;
-      }
-      currentThemeRef.current = theme;
-      setCurrentTheme(theme);
-      if (document.body.dataset.theme !== theme) {
-        document.body.dataset.theme = theme;
-      }
-    };
-
-    updateTheme();
-
-    const observer = new MutationObserver(updateTheme);
-    const targets: (globalThis.Element | null)[] = [
-      document.documentElement,
-      document.body,
-      document.querySelector(".vscode-webview"),
-    ];
-
-    targets
-      .filter((target): target is globalThis.Element => !!target)
-      .forEach((target) =>
-        observer.observe(target, {
-          attributes: true,
-          attributeFilter: ["data-theme"],
-        }),
-      );
-
-    return () => {
-      observer.disconnect();
-    };
-  }, []);
-
-  const getCssVariableValue = useCallback(
-    (variableName: string, fallback: string) => {
-      if (typeof window === "undefined") return fallback;
-      const rootElement =
-        document.querySelector(".vscode-webview") || document.documentElement;
-      const value =
-        getComputedStyle(rootElement).getPropertyValue(variableName);
-      return value?.trim() || fallback;
-    },
-    [],
-  );
-
-  const getMinimapNodeColor = useCallback(
-    (node: Node<ChainGraphNodeData>) => {
-      if (node.type === "swimlane") {
-        return getSwimlaneBorderColor(
-          (node.data.properties as Record<string, unknown>)["color"] as string,
-        );
-      }
-
-      if (node.type === "container") {
-        return getCssVariableValue("--container-header-background", "#fff9e6");
-      }
-
-      if (node.style?.backgroundColor) {
-        return node.style.backgroundColor;
-      }
-
-      if (node.data?.elementType && libraryElements) {
-        const libraryElement = libraryElements.find(
-          (el) => el.name === node.data.elementType,
-        );
-        return getElementColor(libraryElement);
-      }
-
-      return "#fdf39d";
-    },
-    [libraryElements, getCssVariableValue],
-  );
-
-  const getMinimapNodeStrokeColor = useCallback(
-    (node?: Node<ChainGraphNodeData>) => {
-      if (node?.type === "container") {
-        return getCssVariableValue(
-          "--vscode-foreground",
-          currentTheme === "dark"
-            ? "rgba(255, 255, 255, 0.85)"
-            : "rgba(0, 0, 0, 0.88)",
-        );
-      }
-      return getCssVariableValue(
-        "--vscode-border",
-        currentTheme === "dark" ? "#303030" : "#d9d9d9",
-      );
-    },
-    [currentTheme, getCssVariableValue],
-  );
-
   return (
     <Flex className={styles["graph-wrapper"]}>
       <ElementFocusContext.Provider value={fitViewToElementIdRef}>
@@ -647,101 +314,34 @@ const ChainGraphInner: React.FC = () => {
             </>
           )}
         </Require>
-        <div className="react-flow-container" ref={reactFlowWrapper}>
-          <ElkDirectionContextProvider
-            elkDirectionControl={{
-              direction,
-              toggleDirection,
-            }}
-          >
-            <ReactFlow
-              nodes={
-                readOnly
-                  ? nodes.map((node) => ({
-                      ...node,
-                      draggable: false,
-                      connectable: false,
-                    }))
-                  : nodes
-              }
-              nodeTypes={nodeTypes}
-              defaultEdgeOptions={{ zIndex: 1001 }}
-              edges={renderEdges}
-              onNodeDragStart={readOnly ? undefined : onNodeDragStart}
-              onNodeDrag={readOnly ? undefined : onNodeDrag}
-              onNodeDragStop={
-                readOnly
-                  ? undefined
-                  : (event, draggedNode) =>
-                      void onNodeDragStop(event, draggedNode)
-              }
-              onNodesChange={readOnly ? undefined : onNodesChange}
-              onEdgesChange={
-                readOnly ? undefined : (changes) => onEdgesChange(changes)
-              }
-              onConnect={
-                readOnly
-                  ? undefined
-                  : (connection) => void onConnect(connection)
-              }
-              onDelete={
-                readOnly
-                  ? undefined
-                  : (changes) => {
-                      void handleDelete(changes);
-                    }
-              }
-              onDrop={readOnly ? undefined : (event) => void onDrop(event)}
-              onDragOver={readOnly ? undefined : onDragOver}
-              onNodeDoubleClick={readOnly ? undefined : onNodeDoubleClick}
-              zoomOnDoubleClick={false}
-              deleteKeyCode={deleteKeyCode}
-              proOptions={{ hideAttribution: true }}
-              onContextMenu={readOnly ? undefined : onContextMenu}
-              onNodeContextMenu={readOnly ? undefined : onNodeContextMenu}
-              onPaneClick={closeMenu}
-              fitView
-            >
-              <ElementFocus />
-              <Background variant={BackgroundVariant.Dots} />
-              <MiniMap
-                zoomable
-                pannable
-                position="top-right"
-                nodeColor={getMinimapNodeColor}
-                nodeStrokeColor={getMinimapNodeStrokeColor}
-                nodeStrokeWidth={2}
-              />
-              <ChainGraphViewControls
-                before={
-                  <>
-                    {!readOnly && (
-                      <Button
-                        className={controlStyles.button}
-                        type="text"
-                        title="Left Panel"
-                        data-active={leftPanelVisible}
-                        onClick={() => setLeftPanelVisible((prev) => !prev)}
-                        icon={<OverridableIcon name="leftPanel" />}
-                      />
-                    )}
-                    <Button
-                      className={controlStyles.button}
-                      type="text"
-                      title="Right Panel"
-                      data-active={rightPanelVisible}
-                      onClick={() => setRightPanelVisible((prev) => !prev)}
-                      icon={<OverridableIcon name="rightPanel" />}
-                    />
-                  </>
-                }
-                onExpandAllContainers={expandAllContainers}
-                onCollapseAllContainers={collapseAllContainers}
-              />
-              {menu && <ContextMenu menu={menu} closeMenu={closeMenu} />}
-            </ReactFlow>
-          </ElkDirectionContextProvider>
-        </div>
+        <ChainGraphView
+          readOnly={readOnly}
+          submitOpenElement={openElementModal}
+          controls={{
+            before: (
+              <>
+                {!readOnly && (
+                  <Button
+                    className={controlStyles.button}
+                    type="text"
+                    title="Left Panel"
+                    data-active={leftPanelVisible}
+                    onClick={() => setLeftPanelVisible((prev) => !prev)}
+                    icon={<OverridableIcon name="leftPanel" />}
+                  />
+                )}
+                <Button
+                  className={controlStyles.button}
+                  type="text"
+                  title="Right Panel"
+                  data-active={rightPanelVisible}
+                  onClick={() => setRightPanelVisible((prev) => !prev)}
+                  icon={<OverridableIcon name="rightPanel" />}
+                />
+              </>
+            ),
+          }}
+        />
         {rightPanelVisible && (
           <>
             <PanelResizeHandle
